+++ examples/corporate-conundrum/New Game.sql
@@ -1,7 +1,5 @@
 INSERT INTO games(id)
 VALUES(random())
 RETURNING
-    'http_header' as component,
-    'game.sql?id='||id as "Location";
-
-SELECT 'text' as component, 'redirecting to game...' as contents;
+    'redirect' as component,
+    CONCAT('game.sql?id=', id) as link;