+++ tests/sql_test_files/it_works_query_string.sql
@@ -1,0 +1,9 @@
+SELECT
+    'text' AS component,
+    CASE sqlpage.link('', sqlpage.variables('get'))
+        WHEN '?x=1' THEN
+            'It works !'
+        ELSE
+            'Expected "?x=1"'
+    END AS contents
+;