+++ tests/sql_test_files/it_works_run_sql_from_database.sql
@@ -1,7 +1,14 @@
-select 'dynamic' as component,
-    sqlpage.run_sql('tests/display_text.sql', CONCAT('{"html": "', html,'"}')) as properties
-from (
-    select 'It ' as html
-    union all
-    select 'works !'
-) as t1;
+select
+    'dynamic' as component,
+    sqlpage.run_sql (
+        'tests/display_text.sql',
+        CONCAT ('{"html":"', html, '"}') -- UNSAFE. Don't do that with untrusted data. We do it here because we can't use json_object (the syntax is not the same in all supported databases)
+    ) as properties
+from
+    (
+        select
+            'It ' as html
+        union all
+        select
+            'works !'
+    ) as t1;